3.0_flink/requirements.txt
apache-flink==1.17.1
redis-py-cluster==2.1.3
cassandra-driver==3.25.0
kafka-python==2.0.2
uuid==1.30
avro-python3>=1.8.1,<1.10.0
pytz==2023.3
protobuf>=3.19.0,<=3.21.0
python-dateutil==2.8.2 